+++ packages/info/.gitignore
@@ -1,1 +1,2 @@
 *.js
+lib/
