+++ packages/info/package.json
@@ -23,5 +23,8 @@
   "scripts": {
     "build": "tsc",
     "watch": "npm run build && tsc -w"
-  }
+  },
+  "files": [
+    "lib"
+  ]
 }
