+++ packages/info/src/commands.ts
