+++ packages/info/src/configParser.ts
