+++ packages/info/src/index.ts
