+++ packages/info/src/options.ts
