+++ packages/info/src/renderTable.ts
